// src/types/product.ts
export interface Product {
  id: number;
  integrationProductId: string;
  name: string;
  referenceCode: string;
  definition: string;
  brand: string;
  mainCategory: string;
  category: string;
  showOnSite: boolean;
  searchable: boolean;
  showPrice: boolean;
  showAvailability: boolean;
  showStock: boolean;
  shortDescription: string;
  longDescription: string;
  pageTitle: string;
  slug: string;
  metaDescription: string;
  searchTerms: string;
  material: string;
  warranty: string;
  side: string;
  productWeight: number;
  productWidth: number;
  productHeight: number;
  productLength: number;
  commercialCode: string;
  voltage: string;
  application: string;
  vehicleModel: string;
  compatibleVehicles: string;
  sku: {
    id: number;
    integrationId: string;
    name: string;
    referenceCode: string;
    supplier: string;
    ean: string;
    stock: number;
    sellWithoutStock: boolean;
    maxBackorder: number;
    shippingDaysWithStock: number;
    shippingDaysWithoutStock: number;
    minimumStockEnabled: boolean;
    minimumStock: number;
    basePrice: number;
    costPrice: number;
    tax: number;
    promotion: boolean;
    promotionPrice: number;
    virtual: boolean;
    weight: number;
    width: number;
    height: number;
    depth: number;
    condition: string;
    showCondition: boolean;
    packageQuantity: number;
    unitOfMeasure: string;
    pointsProgram: boolean;
    points: number;
  };
}